backend/requirements.txt
# FastAPI y servidor
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
email-validator==2.1.0

# Base de datos
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Almacenamiento
minio==7.2.3

# IA - LLM
requests==2.31.0
httpx==0.26.0

# IA - Embeddings
transformers==4.37.2
torch==2.1.2
sentence-transformers==2.3.1

# IA - Audio/Video
openai-whisper==20231117
ffmpeg-python==0.2.0

# Utilidades
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
validators==0.22.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3